adjusted common/richText and msg/enhance per Ornicar's request
